refactor: share the real_app fixture
Both route tests built the same app off API_BLUEPRINTS and
INTERNAL_BLUEPRINTS, seventeen identical lines each. One fixture in
conftest now, importing the route tree inside the body: it costs
~1.1s and only these two tests need it.
